feat: Add Data for Card in Manage Stock in Admin

+++ dto/response/analytics_response.go
@@ -16,5 +16,5 @@ type AnalyticsResponse struct {
 type DataForManageStockAdmin struct {
 	TotalProduct   uint `json:"totalProduct"`
 	TotalProvider  uint `json:"totalProvider"`
-	CashoutBalance uint `json:"cashoutBalance"`
+	CashoutBalance float64 `json:"cashoutBalance"`
 }
